修正jsonUtils。 透過dao存取的saveEntity(cashDetailEntity)的時候會沒有辦法順利存到資料庫，會跑出null pointer exception的問題，先不解決，直接改為使用repository存資料庫就不會有問題了。
git-svn-id: http://svn.gwis.com.tw:88/repos/M_Charge_Project/trunk@5319 4b8841f4-42b5-4263-839e-868dbe4382dd

+++ src/main/java/com/gateweb/charge/repository/BillCycleRepository.java
@@ -40,7 +40,7 @@ public interface BillCycleRepository extends JpaRepository<BillCycleEntity, Inte
 		java.lang.Integer billId 
 	) throws DaoSystemException;
 
-	public BillCycleEntity findByBillId(Long billId);
+	public BillCycleEntity findByBillId(Integer billId);
 	
 	/*public Page<BillCycleEntity> findByBillId(Long billId, Pageable pageable);
